+++ spec/spec_init_model.rb
@@ -3,34 +3,28 @@
 I18n.available_locales = %i[de en es]
 
 def prepare_database!
-  db = 'translateable_test_db'.freeze
-
-  ActiveRecord::Base.establish_connection(adapter: 'postgresql',
-                                          database: 'template1',
-                                          username: 'postgres')
-
-  begin
-    ActiveRecord::Base.connection.drop_database(db)
-  rescue ActiveRecord::StatementInvalid
-  end
+  db = 'translatable_test_db'
+  connect_db('template1')
+  ActiveRecord::Base.connection.drop_database(db) rescue nil
   ActiveRecord::Base.connection.create_database(db)
+  connect_db(db)
+  ActiveRecord::Base.connection.drop_table :test_models rescue nil
+  migrate!
+end
 
+def connect_db(db)
   ActiveRecord::Base.establish_connection(adapter: 'postgresql',
                                           database: db,
-                                          username: 'postgres')
-
-  begin
-    ActiveRecord::Base.connection.drop_table :test_models
-  rescue ActiveRecord::StatementInvalid
-  end
-
-  migrate!
+                                          username: 'root',
+                                          password: 'password',
+                                          host: 'postgres')
 end
 
 def migrate!
   ActiveRecord::Base.connection.create_table :test_models do |t|
     t.jsonb :title
     t.jsonb :key
+    t.string :title_text, default: '{}'
   end
 end
 
@@ -40,4 +34,5 @@ class TestModel < ActiveRecord::Base
   include BuddyTranslatable
 
   translatable :title, default_key: :de
+  translatable :title_text, default_key: :de
 end