+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0 (17-06-2021)
+- feat: Support for text format attributes (non jsonb attributes)
+- chore: Add container settings
+- refactor: Change default data format from stringified keys into symbolized keys
+- feat: support for locale specific setter
+
 # 0.2.0 (14-08-2020)
 - Removed sales_datable